[IMP] point_of_sale,*: ticket (receipt after payment) revamp
point_of_sale, l10n_co_pos, l10n_fr_pos_cert, pos_loyalty, l10n_in_pos,
l10n_br_edi_pos, pos_self_order, ‎l10n_cl_edi_pos, ‎pos_urban_piper

In this commit:
- Aligned order lines with total, tax line, etc.
- font family changed to `monospace` for prices in ordelines.
- Removed loyalty point lines.
- Adjusted alignment of self-invoicing QR code and loyalty barcode code.
- Added company name and address in the footer, moving contact details from the
  header to the footer.
- Organized preset data properly in the header.
- Added VAT prefix in ticket if county is in `Europian union VAT` country group.
- `Untaxed Amount` text replaced with `Subtotal`.
- Order tracking number configured only to show on self-order or kiosk receipt.

Task: 4552788
