+++ addons/l10n_co_pos/static/src/app/screens/receipt_screen/order_receipt/order_receipt.xml
@@ -2,9 +2,9 @@
 
 <templates id="template" xml:space="preserve">
     <t t-name="l10n_co_pos.OrderReceipt" t-inherit="point_of_sale.OrderReceipt" t-inherit-mode="extension">
-        <xpath expr="//div[hasclass('pos-receipt-order-data')]" position="inside">
+        <xpath expr="//div[hasclass('before-footer')]" position="inside">
             <t t-if="order.name !== false">
-                <div style="word-wrap:break-word;"><t t-esc="order.name"/></div>
+                <div style="word-wrap:break-word;" class="pt-3"><t t-out="order.name"/></div>
             </t>
         </xpath>
     </t>
