+++ addons/l10n_fr_pos_cert/static/src/xml/OrderReceipt.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates id="template" xml:space="preserve">
     <t t-name="l10n_fr_pos_cert.OrderReceipt" t-inherit="point_of_sale.OrderReceipt" t-inherit-mode="extension">
-        <xpath expr="//div[hasclass('pos-receipt-order-data')]" position="inside">
+        <xpath expr="//div[hasclass('before-footer')]" position="inside">
             <t t-if="order.l10n_fr_hash !== false">
-                <br/>
-                <div style="word-wrap:break-word;"><t t-out="order.l10n_fr_hash"/></div>
+                <div class="text-break pt-3"><t t-out="order.l10n_fr_hash"/></div>
             </t>
         </xpath>
     </t>
