+++ addons/point_of_sale/static/src/app/components/order_display/order_display.xml
@@ -13,7 +13,7 @@
                         t-attf-class="{{this.props.mode === 'receipt' ? 'mt-1 bg-opacity-75 p-0 ' : 'customer-note w-100 p-2 text-break bg-warning bg-opacity-25 mt-0 text-warning'}}">
                         <div class="flex-wrap w-100 m-0">
                             <div class="ps-1">
-                                <t t-foreach="order.general_customer_note.split('\n')" t-as="subNote" t-key="subNote_index">
+                                <t t-foreach="order.general_customer_note.trim().split('\n')" t-as="subNote" t-key="subNote_index">
                                     <div class="d-inline text-break">
                                         <t t-esc="subNote"/><br/>
                                     </div>
@@ -21,7 +21,7 @@
                             </div>
                         </div>
                     </div>
-                    <div class="internal-note-container d-flex flex-wrap gap-2 p-2">
+                    <div t-if="order.internal_note" class="internal-note-container d-flex flex-wrap gap-2" t-attf-class="{{this.props.mode === 'receipt' ? 'py-2' : 'p-2'}}">
                         <TagsList tags="getInternalNotes()"/>
                     </div>
                 </div>
