+++ addons/point_of_sale/static/src/app/components/orderline/orderline.xml
@@ -2,47 +2,55 @@
 <templates id="template" xml:space="preserve">
     <t t-name="point_of_sale.Orderline">
         <li t-if="line.order_id" class="orderline position-relative d-flex align-items-center lh-sm cursor-pointer"
-            t-attf-class="{{ line.combo_parent_id ? 'orderline-combo fst-italic ms-4 border-start' : '' }}"
-            t-att-class="{'selected' : line.isSelected() and this.props.mode === 'display', ...line.getDisplayClasses(), ...(props.class || [])}">
+            t-attf-class="{{ line.combo_parent_id ? 'orderline-combo fst-italic ms-4' : '' }}"
+            t-att-class="{'selected' : line.isSelected() and this.props.mode === 'display', ...line.getDisplayClasses(), ...(props.class || []), 'border-start': props.mode != 'receipt' and line.combo_parent_id}">
             <div class="product-order"></div>
             <div t-if="props.showImage and line.product_id.getImageUrl()" class="o_line_container d-flex align-items-center justify-content-center">
                 <img t-attf-style="border-radius: 1rem;" t-att-src="line.product_id.getImageUrl()"/>
             </div>
             <div class="d-flex flex-column w-100"
-                 t-attf-class="{{ line.combo_parent_id ? 'px-2 py-1' : 'p-2' }}">
-                <div class="line-details d-flex justify-content-between align-items-center">
-                    <div class="product-name d-flex flex-grow-1 align-items-center pe-2 text-truncate">
-                        <span class="qty d-inline-block px-1 fw-bolder">
+                 t-attf-class="{{ line.combo_parent_id ? props.mode == 'receipt' ? 'px-2' : 'p-2' : props.mode == 'receipt' ? line.combo_line_ids.length > 0 ? '' : 'py-1' : 'p-2' }}">
+                <div class="line-details d-flex justify-content-between" t-att-class="{ 'align-items-center' : props.mode == 'receipt' }">
+                    <div class="product-name d-flex flex-grow-1 pe-2 text-truncate" t-att-class="{ 'align-items-center' : props.mode == 'receipt' }">
+                        <span class="qty d-inline-block fw-bolder"
+                            t-attf-class="{{ props.mode == 'receipt' ? 'pe-2' : 'px-1' }}">
                             <t t-esc="line.getQuantityStr().unitPart"/>
-                            <small t-if="line.getQuantityStr().decimalPart" class="fw-normal text-muted" t-esc="line.getQuantityStr().decimalPart"/>
-                        </span>
-                        <span class="text-wrap d-inline">
-                            <t t-esc="line.orderDisplayProductName.name" />
-                            <br/>
-                            <small class="attribute-line fst-italic" t-if="line.orderDisplayProductName.attributeString">
-                                - <t t-esc="line.orderDisplayProductName.attributeString" />
+                            <small t-if="line.getQuantityStr().decimalPart" class="fw-normal" style="font-size: 85%;">
+                                <t t-esc="line.getQuantityStr().decimalPoint"/>
+                                <t t-esc="line.getQuantityStr().decimalPart"/>
                             </small>
                         </span>
+                        <span class="text-wrap d-inline">
+                            <t t-if="props.mode == 'receipt'">
+                                <t t-out="line.full_product_name" />
+                            </t>
+                            <t t-else="">
+                                <t t-esc="line.orderDisplayProductName.name" />
+                                <br/>
+                                <small class="attribute-line fst-italic" t-if="line.orderDisplayProductName.attributeString">
+                                    - <t t-esc="line.orderDisplayProductName.attributeString" />
+                                </small>
+                            </t>
+                        </span>
                     </div>
-                    <div t-if="!props.basic_receipt" t-esc="line.getPriceString()" class="product-price price fw-bolder"/>
+                    <div t-if="!props.basic_receipt" t-out="line.getPriceString()" class="product-price price" t-attf-class="{{ props.mode == 'receipt' ? 'font-monospace' : 'fw-bolder' }}"/>
                     <div t-if="props.showTaxGroup" t-esc="this.taxGroup" class="text-end" style="width: 2rem"/>
                 </div>
-                <ul class="info-list d-flex flex-column mb-0" t-att-class="{'gap-1' : line.customer_note || line.note || line.discount || line.packLotLines?.length}">
+                <ul class="info-list d-flex flex-column mb-0 ms-2" t-att-style="props.mode == 'receipt' and 'font-size: 75%;'" t-attf-class="{{ props.mode === 'receipt' ? '' : (line.customer_note || line.note || line.discount || line.packLotLines?.length) ? 'gap-2 mt-1' : '' }}">
+                    <li class="price-per-unit" t-if="!props.basic_receipt and this.props.line.getQuantity() != 1 and (props.mode == 'receipt' || (line.price_type !== 'original' and !line.combo_parent_id))">
+                        <t t-if="line.price !== 0">
+                            <t t-esc="formatCurrency(line.unitDisplayPrice)" />
+                        </t>
+                        /
+                        <t t-if="line.product_id.uom_id?.name" t-esc="line.product_id.uom_id?.name" />
+                    </li>
                     <li t-if="line.discount" class="price-per-unit">
                         <t t-set="discount" t-value="line.getDiscountStr()" />
                         <t t-if="!props.basic_receipt and line.price_unit !== 0 and discount and discount !== '0'">
                             <i class="fa fa-tag pe-1"/><em><t t-esc="discount" />% </em> discount off on <t t-esc="env.utils.formatCurrency(line.allPrices.priceWithTaxBeforeDiscount)"/>
                         </t>
                     </li>
-                    <li class="price-per-unit" t-if="!props.basic_receipt and (props.mode == 'receipt' || (line.price_type !== 'original' and !line.combo_parent_id))">
-                        <t t-if="line.price !== 0">
-                            <i class="fa fa-money center pe-1"/>
-                            <t t-esc="formatCurrency(line.unitDisplayPrice)" />
-                        </t>
-                        /
-                        <t t-if="line.product_id.uom_id?.name" t-esc="line.product_id.uom_id?.name" />
-                    </li>
-                    <li t-if="line.customer_note" class="customer-note w-100 rounded text-break bg-opacity-25 mb-1" t-att-class="{'bg-warning text-warning p-2 mt-2': this.props.mode === 'display'}">
+                    <li t-if="line.customer_note" class="customer-note w-100 rounded text-break bg-opacity-25" t-att-class="{'bg-warning text-warning p-2': this.props.mode === 'display'}">
                         <i class="fa fa-sticky-note me-1" role="img" aria-label="Customer Note" title="Customer Note"/>
                         <t t-esc="line.customer_note" />
                     </li>
@@ -50,7 +58,7 @@
                         <TagsList tags="getInternalNotes()" />
                     </div>
                     <li t-if="line.product_id.tracking !== 'none'" class="row flex-wrap w-100 m-0">
-                        <div class="col-auto px-1">
+                        <div class="col-auto" t-attf-class="{{ props.mode == 'receipt' ? 'p-0' : 'px-1' }}">
                             <t t-slot="pack-lot-icon" />
                         </div>
                         <div class="col ps-0">
