+++ addons/point_of_sale/static/src/app/components/product_card/product_card.scss
@@ -45,7 +45,7 @@
     }
 
     > .qty {
-        min-width: 2rem;
+        min-width: 2.5rem;
     }
 }
