+++ addons/point_of_sale/static/src/css/pos_receipts.css
@@ -1,6 +1,6 @@
 .pos-receipt-print {
     width: 512px;
-    font-size: 27px;
+    font-size: 22px;
     color: #000000;
 }
 
@@ -34,12 +34,10 @@
 }
 
 .pos-receipt .pos-receipt-order-data {
-    text-align: center;
     font-size: 75%;
 }
 
 .pos-receipt .pos-receipt-amount {
-    font-size: 125%;
     text-align: start;
 }
 
@@ -85,4 +83,12 @@
 .pos-receipt .orderlines {
     /*rtl:ignore*/
     direction: ltr;
+}
+
+.pos-receipt-print .qty {
+    min-width: 3.5rem;
+}
+
+.pos-receipt-print .info-list {
+    margin-left: 1.5rem !important;
 }
