+++ addons/pos_loyalty/static/src/app/screens/receipt_screen/order_receipt/order_receipt.xml
@@ -3,55 +3,35 @@
 
     <t t-name="pos_coupon.OrderReceipt" t-inherit="point_of_sale.OrderReceipt" t-inherit-mode="extension">
         <xpath expr="//div[hasclass('pos-receipt')]//div[hasclass('before-footer')]" position="inside">
-            <t t-foreach="order.getLoyaltyPoints() or []" t-as="_loyaltyStat" t-key="_loyaltyStat.couponId">
-                <!-- Show only if portal_visible. -->
-                <div t-if="_loyaltyStat.program.portal_visible and (_loyaltyStat.points.won || _loyaltyStat.points.spent)" class='loyalty'>
-                    <span class="pos-receipt-center-align">
-                        <div>--------------------------------</div>
-                    </span>
-                    <t t-if='_loyaltyStat.points.won'>
-                        <div><t t-esc="_loyaltyStat.points.name"/> Won: <span t-esc='_loyaltyStat.points.won' class="pos-receipt-right-align"/></div>
-                    </t>
-                    <t t-if='_loyaltyStat.points.spent'>
-                        <div><t t-esc="_loyaltyStat.points.name"/> Spent: <span t-esc='_loyaltyStat.points.spent' class="pos-receipt-right-align"/></div>
-                    </t>
-                    <!-- Don't use points.total, it's wrong in this context (after the order synced). -->
-                    <!-- Show balance as it's updated during _postPushOrderResolve. -->
-                    <t t-if='_loyaltyStat.points.balance'>
-                        <div>Balance <t t-esc="_loyaltyStat.points.name"/>: <span t-esc='_loyaltyStat.points.balance' class="pos-receipt-right-align"/></div>
-                    </t>
-                </div>
-            </t>
-            <t t-if="order.partner_id">
-                <br/>
-                <div>Customer <span t-esc='order.partner_id.name' class='pos-receipt-right-align'/></div>
-            </t>
             <t t-if="order.new_coupon_info and order.new_coupon_info.length !== 0">
-                <div class="pos-coupon-rewards">
-                    <div>------------------------</div>
-                    <br/>
-                    <div>
-                        Coupon Codes
-                    </div>
+                <div class="pos-coupon-rewards pt-3">
                     <t t-foreach="order.new_coupon_info" t-as="coupon_info" t-key="coupon_info.code">
+                        <t t-if="order.partner_id and order.preset_id?.identification != 'address'">
+                            <div class="d-flex pb-1" style="font-size: 75%;">
+                                <span>Customer</span>
+                                <span t-out="order.partner_id.name" class="ms-auto fw-bolder"/>
+                            </div>
+                        </t>
                         <div class="coupon-container">
-                            <div style="font-size: 110%;">
-                                <t t-esc="coupon_info['program_name']"/>
-                            </div>
-                            <div>
-                                <span>Valid until: </span> 
-                                <t t-if="coupon_info['expiration_date']">
-                                    <t t-esc="coupon_info['expiration_date']"/>
-                                </t>
-                                <t t-else="">
-                                    no expiration
-                                </t>
-                            </div>
-                            <div>
-                                <img t-att-src="'/report/barcode/Code128/'+coupon_info['code']" style="width:200px;height:50px" alt="Barcode"/>
-                            </div>
-                            <div>
-                                <t t-esc="coupon_info['code']"/>
+                            <div class="row g-2">
+                                <div class="col-6">
+                                    <t t-out="coupon_info['program_name']"/>
+                                </div>
+                                <div class="col-6">
+                                    <img style="transform: translateX(13%);" t-att-src="'/report/barcode/Code128/'+coupon_info['code']" class="img-fluid" alt="Barcode"/>
+                                </div>
+                                <div class="col-6">
+                                    <span>Until: </span>
+                                    <t t-if="coupon_info['expiration_date']">
+                                        <t t-out="coupon_info['expiration_date']"/>
+                                    </t>
+                                    <t t-else="">
+                                        no expiration
+                                    </t>
+                                </div>
+                                <div class="col-6 text-end">
+                                    <t t-out="coupon_info['code']"/>
+                                </div>
                             </div>
                         </div>
                     </t>
