+++ addons/pos_loyalty/static/src/css/Loyalty.scss
@@ -1,9 +1,3 @@
-.pos-receipt .pos-coupon-rewards {
-    text-align: center;
-    padding: 1em;
-}
-
 .pos-coupon-rewards .coupon-container {
-    padding-top: 1em;
     font-size: 75%;
 }
