+++ addons/pos_self_order/static/src/app/self_order_index.scss
@@ -85,3 +85,7 @@ li {
 .touch-device * {
     cursor: none !important;
 }
+
+.pos-receipt .info-list {
+    padding-left: 2rem;
+}
