+++ addons/pos_self_order/static/src/overrides/components/receipt_header/receipt_header.xml
@@ -2,17 +2,18 @@
 <templates id="template" xml:space="preserve">
     <t t-name="pos_self_order.ReceiptHeader" t-inherit="point_of_sale.ReceiptHeader" t-inherit-mode="extension">
         <xpath expr="//div[hasclass('pos-receipt-contact')]" position="before">
-            <div t-if="order.preset_id or order.config.self_ordering_mode !== 'nothing'" class="picking-service text-center pb-2">
-                <span t-if="order.preset_id?.identification == 'none' or (!order.preset_id and order.config.self_ordering_service_mode == 'table')">Service at Table</span>
-                <span t-elif="order.preset_id?.identification == 'name' or (!order.preset_id and order.config.self_ordering_service_mode != 'table')">Pickup At Counter</span>
-                <span t-else="">Delivery</span>
-            </div>
-        </xpath>
-        <xpath expr="//div[hasclass('pos-receipt-contact')]" position="before">
-            <div t-if="order.config.self_ordering_mode !== 'nothing' and order.table_stand_number" class="table-tracker text-center pb-2">
-                Table Tracker:
-                <span class="pt-3" t-esc="order.table_stand_number" />
-            </div>
+            <t t-if="order.session?._self_order_pos">
+                <div t-if="order.preset_id or order.config.self_ordering_mode !== 'nothing'" class="picking-service text-center pt-3">
+                    <span t-if="order.preset_id?.identification == 'none' or (!order.preset_id and order.config.self_ordering_service_mode == 'table')">Service at Table</span>
+                    <span t-elif="order.preset_id?.identification == 'name' or (!order.preset_id and order.config.self_ordering_service_mode != 'table')">Pickup At Counter</span>
+                    <span t-else="">Delivery</span>
+                </div>
+                <h1 class="tracking-number text-center" style="font-size: 100px" t-out="order.tracking_number" />
+                <div t-if="order.config.self_ordering_mode !== 'nothing' and order.table_stand_number" class="table-tracker text-center">
+                    Table Tracker:
+                    <span class="pt-3" t-out="order.table_stand_number" />
+                </div>
+            </t>
         </xpath>
     </t>
 </templates>
